backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient>=2.0
python-dotenv==1.0.1
django-cors-headers
Pillow>=11.2.1
whitenoise==6.2.0
drf-spectacular[sidecar]
